requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
dill==0.3.8
re
nltk==3.8.1
scikit-learn==1.4.2
python-pptx==0.6.23
PyPDF2==3.0.1
wordcloud==1.9.4 
Flask==3.0.3
dill==0.3.4
pandas==1.3.3
chardet==4.0.0
PyPDF2==1.26.0
python-docx==0.8.11
scikit-learn==0.24.2